+++ src/spice2x/games/popn/bi3a_hook.h
@@ -3,7 +3,7 @@
 #if SPICE64
 
 namespace games::popn {
-    void bi2x_hook_init();
+    void bi3a_hook_init();
 }
 
 #endif
